+++ src/main/java/com/dashur/integration/extw/connectors/relaxgaming/data/service/GetFeatureTriggersRequest.java
@@ -20,7 +20,7 @@ import lombok.ToString;
 public class GetFeatureTriggersRequest extends ServiceRequest {
 
   @JsonProperty("playerid")
-  private Integer playerId;
+  private Long playerId;
 
   @JsonProperty("partnerid")
   private Integer partnerId;
